+++ tests/file.ts
@@ -1,4 +1,4 @@
-import { mergician } from '../dist/mergician.mjs';
+import { mergician } from '../dist/mergician.esm.js';
 
 // Use for JSDoc / IntelliSense testing
 mergician;