+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `hoistProto` option
 - Add `skipProto` option
 - Add TypeScript support via type declarations
+- Update JSDoc comments for improved code hinting
 
 ## 1.1.0